% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_dup.R
\name{clean_dup}
\alias{clean_dup}
\title{Function to clean duplicated longitude and latitude data}
\usage{
clean_dup(data, longitude, latitude, threshold = 0)
}
\arguments{
\item{data}{A data.frame with longitude and latitude data}

\item{longitude}{A character vector of the column name of longitude.}

\item{latitude}{A character vector of the column name of latitude.}

\item{threshold}{A numerc value representig the eculidean distance between coordinates
to be considered as a duplicate.}
}
\value{
Returns a data.frame with coordinate data from species
}
\description{
Clean duplicated longitude and latitude data using threshold distance
             which is a distance between points to be considered
             duplicates.
}
\examples{
occ_rec <- read.csv(system.file("exdata/cardon_virtual.csv",
                                 package = "hsi"))
occ_clean <- hsi::clean_dup(occ_rec,longitude="x",
                            latitude="y",threshold=0.5)
}
