% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_dup_by_year_df.R
\name{clean_dup_by_year_df}
\alias{clean_dup_by_year_df}
\title{clean_dup_by_year_df: Function to clean duplicated ocurrence data by year.}
\usage{
clean_dup_by_year_df(df_coords_year, longitude, latitude, year_var, threshold)
}
\arguments{
\item{df_coords_year, }{A data.frame with coordinate records by year.}

\item{longitude}{A character vector of the column name of longitude.}

\item{latitude}{A character vector of the column name of latitude.}

\item{year_var}{The name of the year variable.}

\item{threshold}{A numerc value representig the euclidean distance between coordinates.}
}
\description{
Clean duplicated occurence data by year using a threshold distance.
}
\examples{
puma_recs <- read.csv(system.file("exdata/puma_coords_year.csv",
                                 package = "hsi"))
puma_clean_year <- clean_dup_by_year_df(df_coords_year = puma_recs,
                                     longitude = "longitude",
                                     latitude = "latitude",
                                     year_var = "year",
                                     threshold = 0.05)
}
\seealso{
Use this version of the function \code{\link[hsi]{clean_dup_by_year}} for data.frames
}
