% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_center.R
\name{cov_center}
\alias{cov_center}
\title{Function to compute the Minimum Volume covariance Matrix of an ellipsoid niche model.}
\usage{
cov_center(data, mve = TRUE, level, vars = NULL)
}
\arguments{
\item{data}{A data.frame or a matrix with the numeric values of the variables
that will be used to model the niche.}

\item{mve}{A logical value. If TRUE a minimum volume ellipsoid will be computed using
the function \code{\link[MASS]{cov.mve}} of the \pkg{MASS} package. If False the covariance matrix of the input data will be used.}

\item{level}{A numerical value specifying the proportion of the data to be
used to compute the ellipsoid.}

\item{vars}{A numeric or a string vector specifying the columns indexes/names of the variables of the input data which will be used to fit the ellipsoid model. If NULL the user will be asked to enter the indexes.
interactively}
}
\value{
Returns a list containing the centroid of the ellipsoid, the covariance matrix based on
the input data, ellipsoid volume, semi-axis length and axis coordinates.
}
\description{
Function to compute the covariance matrix, the niche centroid and volume of an
ellipsoid model. It uses the values of the niche variables of the ocurrences points.
}
\examples{
\dontrun{
cardon_data <- read.csv(system.file("exdata/cardon_virtual.csv",
                                    package = "hsi"))
cardon_ellipsoid <- cov_center(data=cardon_data, mve=TRUE, level=0.975,
                               vars=c("bio5","bio6", "bio12"))
# To obtain the coordinates of the ellipsoid's centroid run:
print(cardon_ellipsoid$centroid)

# To obtain the coordinates of the ellipsoid's minimum value covariance run:
print(cardon_ellipsoid$covariance)

# To obtain the coordinates of the ellipsoid's niche volume run:
print(cardon_ellipsoid$niche_volume)

# To obtain the coordinates of the ellipsoid's Semi Axis Length run:
print(cardon_ellipsoid$SemiAxis_length)

# To obtain the coordinates of the ellipsoid's coordinates run:
print(cardon_ellipsoid$axis_coordinates)

# To plot the ellipsoid with a 99\% of data conteined from the obtained
# by minimim volume method
library(rgl)
ellipsoid_coor <- rgl::ellipse3d(x=cardon_ellipsoid$covariance,centre=cardon_ellipsoid$centroid,
                                 level=0.99)

rgl::plot3d(ellipsoid_coor, alpha=0.3)

}
}
