% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_best_model_ntbox.R
\name{find_best_model_ntbox}
\alias{find_best_model_ntbox}
\title{Function to find the best n-dimensional ellipsoid model using Partial Roc as a performance criteria.}
\usage{
find_best_model_ntbox(
  this_species,
  cor_threshold = 0.9,
  nbg_points = 50000,
  omr_criteria = 0.1,
  ellipsoid_level = 0.975,
  nvars_to_fit = c(2, 3),
  E = 0.05,
  RandomPercent = 50,
  NoOfIteration = 1000,
  parallel = TRUE,
  n_cores = 6
)
}
\arguments{
\item{this_species, }{Species Temporal Environment "sp.temporal.env" object see \code{\link[hsi]{extract_by_year}}.}

\item{cor_threshold}{Threshold valuefrom which it is considered that the correlation is high see \code{\link[hsi]{correlation_finder}}.}

\item{nbg_points}{Number of background points used to compute partial ROC test. See \code{\link[ntbox]{sample_envbg}} for more details.}

\item{omr_criteria}{Omission rate used to select best models. See \code{\link[ntbox]{ellipsoid_selection}} for more details.}

\item{ellipsoid_level}{The proportion of points to be included inside the ellipsoid see \code{\link[hsi]{ellipsoidfit}}.}

\item{nvars_to_fit}{Number of variables that will be used to model.}

\item{E}{Amount of error admissible for Partial Roc test (by default =.05). Value should range between 0 - 1. see \code{\link[hsi]{PartialROC}}}

\item{RandomPercent}{Occurrence points to be sampled in randomly for the boostrap of the Partial Roc test \code{\link[hsi]{PartialROC}}.}

\item{NoOfIteration}{Number of iteration for the bootstrapping of the Partial Roc test \code{\link[hsi]{PartialROC}}.}

\item{parallel}{Logical argument to run computations in parallel. Default TRUE}

\item{n_cores}{Number of cores to be used in parallelization. Default 4}
}
\value{
A "sp.temp.best.model" object with metadata of the best model given the performance of the Partial Roc test.
}
\description{
Function to find the best n-dimensional ellipsoid model using Partial Roc as a performance criteria.
}
