% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsi-package.R
\docType{package}
\name{hsi-package}
\alias{hsi}
\alias{hsi-package}
\title{hsi: Manipulate Habitat Suitability Indices}
\description{
Manipulates Habitat Suitability Indices.
}
\author{
\strong{Maintainer}: Joe Thorley \email{joe@poissonconsulting.ca} (0000-0002-7683-4592)

}
\keyword{internal}
