% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by.R
\name{hsi_by}
\alias{hsi_by}
\title{HSI By}
\usage{
hsi_by(x)
}
\arguments{
\item{x}{A numeric vector of habitat values.}
}
\value{
A number of the habitat increments.
}
\description{
HSI By
}
\examples{
hsi_by(c(1, 1.5, 2))
}
