% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hsi_data}
\alias{hsi_data}
\title{HSI Data}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 10 rows and 2 columns.}
\usage{
hsi_data
}
\description{
HSI Data
}
\examples{
hsi_data
}
\keyword{datasets}
