% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{hsi_plot}
\alias{hsi_plot}
\title{Plot HSI Data}
\usage{
hsi_plot(x, habitat = "Habitat", index = "Index")
}
\arguments{
\item{x}{A data frame with}

\item{habitat}{A string of the name of the column with habitat values.}

\item{index}{A string of the name of the column with index values.}
}
\value{
A ggplot object
}
\description{
Plot HSI Data
}
\examples{
hsi_plot(hsi_data)
}
