% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pref.R
\name{hsi_pref}
\alias{hsi_pref}
\title{HSI Use}
\usage{
hsi_pref(use, avail, habitat = "Habitat", index = "Index")
}
\arguments{
\item{use}{A hsi data frame of use}

\item{avail}{A hsi data frame of availability}

\item{habitat}{A string of the name of the column with habitat values.}

\item{index}{A string of the name of the column with index values.}
}
\value{
A hsi data frame of preference for overlapping habitats.
}
\description{
HSI Use
}
\examples{
hsi_pref(hsi_data, hsi_data)
}
