% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by.R
\name{hsi_seq_by}
\alias{hsi_seq_by}
\title{HSI Sequence}
\usage{
hsi_seq_by(x, by = hsi_by(x))
}
\arguments{
\item{x}{A vector of habitat values}

\item{by}{A number of the increments.}
}
\value{
A vector of the sequence
}
\description{
HSI Sequence
}
\examples{
hsi_seq_by(c(1, 2, 1.5))
}
