% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by.R
\name{hsi_set_by}
\alias{hsi_set_by}
\title{HSI By}
\usage{
hsi_set_by(x, habitat = "Habitat", index = "Index",
  by = hsi_by(x[[habitat]]))
}
\arguments{
\item{x}{A HSI data frame.}

\item{habitat}{A string of the name of the column with habitat values.}

\item{index}{A string of the name of the column with index values.}

\item{by}{A non-negative number of the habitat increments.}
}
\value{
The modified HSI data frame.
}
\description{
HSI By
}
\examples{
hsi_by(c(1, 1.5, 2))
}
