% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplier.R
\name{hsi_set_multiplier}
\alias{hsi_set_multiplier}
\title{HSI Multiplier}
\usage{
hsi_set_multiplier(x, habitat = "Habitat", index = "Index",
  hsi_multi = hsi_multiplier(x))
}
\arguments{
\item{x}{A HSI object.}
}
\value{
A double scalar of the hsi multiplier.
}
\description{
HSI Multiplier
}
