% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transect.R
\name{hsi_transect_set_by}
\alias{hsi_transect_set_by}
\title{Transect Set By}
\usage{
hsi_transect_set_by(x, distance = "Distance", habitat = "Habitat",
  by = hsi_by(x[[distance]]))
}
\arguments{
\item{x}{The object to check.}

\item{distance}{A string of the name of the column with distance values.}

\item{habitat}{A string of the name of the column with habitat values}

\item{by}{A non-negative number of the habitat increments.}
}
\value{
A HSI data.
}
\description{
Transect Set By
}
\examples{
hsi_transect_set_by(trans_data)
}
