% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transect.R
\name{hsi_transect_to_index}
\alias{hsi_transect_to_index}
\title{Transect to HSI Index}
\usage{
hsi_transect_to_index(x, distance = "Distance", habitat = "Habitat",
  hsi_multi = 1, n = 10^6)
}
\arguments{
\item{x}{A data frame.}

\item{distance}{A string of the column name with transect distance values.}

\item{habitat}{A string of the column name with habitat values.}

\item{n}{A count of the number of samples to take.}
}
\value{
A hsi data frame
}
\description{
Transect to HSI Index
}
\examples{
hsi_transect_to_index(trans_data)
}
