% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transect.R
\name{hsi_transect_to_sample}
\alias{hsi_transect_to_sample}
\title{Transect to Sample}
\usage{
hsi_transect_to_sample(x, distance = "Distance", habitat = "Habitat",
  n = 10^6)
}
\arguments{
\item{x}{A data frame.}

\item{distance}{A string of the column name with transect distance values.}

\item{habitat}{A string of the column name with habitat values.}

\item{n}{A count of the number of samples to take.}
}
\value{
A vector of samples.
}
\description{
Transect to Sample
}
\examples{
hsi_transect_to_sample(trans_data, n = 10)
}
