% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use.R
\name{hsi_use}
\alias{hsi_use}
\title{HSI Use}
\usage{
hsi_use(pref, avail, habitat = "Habitat", index = "Index")
}
\arguments{
\item{pref}{A hsi data frame of preference}

\item{avail}{A hsi data frame of availability}

\item{habitat}{A string of the name of the column with habitat values.}

\item{index}{A string of the name of the column with index values.}
}
\value{
A hsi data frame of use for overlapping habitats.
}
\description{
HSI Use
}
\examples{
hsi_use(hsi_data, hsi_data)
}
