% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trans_data}
\alias{trans_data}
\title{Transect Data}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6 rows and 2 columns.}
\usage{
trans_data
}
\description{
Transect Data
}
\examples{
trans_data
}
\keyword{datasets}
