% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_hsi}
\alias{check_hsi}
\title{Check HSI Data}
\usage{
check_hsi(x, habitat = "Habitat", index = "Index",
  hsi_multi = hsi_multiplier(x), unique = TRUE, sorted = unique,
  x_name = substitute(x))
}
\arguments{
\item{x}{The object to check.}

\item{habitat}{A string of the name of the column with habitat values.}

\item{index}{A string of the name of the column with index values.}

\item{unique}{A flag indicating whether the habitat values must be unique.}

\item{sorted}{A flag indicating whether the habitat values must be sorted.}

\item{x_name}{A string of the name of the object.}
}
\value{
An invisible copy of the original object.
}
\description{
A HSI data frame is a data frame with an integer Habitat column 
and a double Index column and an optional attribute hsi_multiplier 
indicating the habitat multiplier.
}
\examples{
check_hsi(hsi_data)
}
