% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_hsi}
\alias{check_hsi}
\title{Check HSI Data}
\usage{
check_hsi(x, habitat = "Habitat", index = "Index",
  by = hsi_by(x[[habitat]]), x_name = substitute(x))
}
\arguments{
\item{x}{The object to check.}

\item{habitat}{A string of the name of the column with habitat values.}

\item{index}{A string of the name of the column with index values.}

\item{by}{A non-negative number of the habitat increments.
If by = NULL then does not check habitat increments}

\item{x_name}{A string of the name of the object.}
}
\value{
An invisible copy of the original object.
}
\description{
A HSI data frame is a data frame with a column of habitat and index values.
}
\examples{
check_hsi(hsi_data)
}
