% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_transect}
\alias{check_transect}
\title{Check Transect}
\usage{
check_transect(x, distance = "Distance", habitat = "Habitat",
  x_name = substitute(x))
}
\arguments{
\item{x}{The object to check.}

\item{distance}{A string of the name of the column with distance values.}

\item{habitat}{A string of the name of the column with habitat values}

\item{x_name}{A string of the name of the object.}
}
\value{
An invisible copy of the original object.
}
\description{
Check Transect
}
\examples{
check_transect(trans_data)
}
