% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{hsi_sample_to_index}
\alias{hsi_sample_to_index}
\title{Sample to Index}
\usage{
hsi_sample_to_index(x, hsi_multi = 1)
}
\arguments{
\item{x}{A double vector of habitat values.}

\item{hsi_multi}{The hsi multiplier for the resultant object}
}
\value{
A hsi data frame with columns Habitat and Index.
}
\description{
Sample to Index
}
\examples{
hsi_sample_to_index(runif(100, 1, 2))
hsi_sample_to_index(runif(100, 1, 2), hsi_multi = 0.1)
}
