% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{hsi_sample_to_index}
\alias{hsi_sample_to_index}
\title{Sample to Index}
\usage{
hsi_sample_to_index(x, by = hsi_by(x))
}
\arguments{
\item{x}{A double vector of habitat values.}

\item{by}{The increments for the hsi Habitat values.}
}
\value{
A hsi data frame with columns Habitat and Index.
}
\description{
Sample to Index
}
\examples{
hsi_sample_to_index(runif(100, 1, 2), by = 0.1)
}
