% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hsi_seq_by}
\alias{hsi_seq_by}
\title{HSI Seq By}
\usage{
hsi_seq_by(x, by = hsi_by(x))
}
\arguments{
\item{x}{A numeric vector of habitat values.}

\item{by}{A non-negative number of the habitat increments.}
}
\value{
A numeric vector of equal spaced values spanning the range of x.
}
\description{
HSI Seq By
}
\examples{
hsi_seq_by(c(1,2,1.5))
hsi_seq_by(c(1,2,1.5), by = 1)
}
