% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by.R
\name{hsi_set_by}
\alias{hsi_set_by}
\title{HSI Set By}
\usage{
hsi_set_by(x, habitat = "Habitat", index = "Index", by = hsi_by(x[[habitat]]))
}
\arguments{
\item{x}{A HSI data frame.}

\item{habitat}{A string of the name of the column with habitat values.}

\item{index}{A string of the name of the column with index values.}

\item{by}{A number of the increments.}
}
\value{
A HSI data.
}
\description{
HSI Set By
}
\examples{
hsi_set_by(hsi_data, by = 2)
}
